Refactored MetricEvolution into separate files
