+++ metric_learn/evolution/__init__.py
@@ -1,0 +1,1 @@
+from .evolution import MetricEvolution
