+++ metric_learn/evolution/fitness/__init__.py
@@ -1,0 +1,5 @@
+from .class_separation import ClassSeparationFitness
+from .classifier import ClassifierFitness
+from .random import RandomFitness
+from .wfme import WeightedFMeasureFitness
+from .wpurity import WeightedPurityFitness
