+++ metric_learn/evolution/strategy/__init__.py
@@ -1,0 +1,5 @@
+from .base import BaseEvolutionStrategy
+from .cmaes import CMAES
+from .dde import DynamicDifferentialEvolution
+from .de import DifferentialEvolution
+from .jde import SelfAdaptingDifferentialEvolution
