+++ metric_learn/evolution/strategy/jde.py
@@ -1,0 +1,99 @@
+from deap import tools
+
+import numpy as np
+
+from .base import BaseEvolutionStrategy
+
+
+class SelfAdaptingDifferentialEvolution(BaseEvolutionStrategy):
+    def __init__(self, population_size=None, Fl=0.1, Fu=0.9, t1=0.1, t2=0.1,
+                 **kwargs):
+        super().__init__(**kwargs)
+
+        self.params.update({
+            'population_size': population_size,
+            'Fl': Fl,
+            'Fu': Fu,
+            't1': t1,
+            't2': t2,
+        })
+
+    def cut_individual(self, individual):
+        return individual[2:]
+
+    def best_individual(self):
+        return self.cut_individual(self.hall_of_fame[0])
+
+    def fit(self, X, y):
+        individual_size = self.params['n_dim'] + 2
+
+        toolbox = self.create_toolbox()
+        toolbox.register("attr_float", np.random.uniform, -1, 1)
+        toolbox.register(
+            "individual",
+            self.generate_individual_with_fitness,
+            toolbox.attr_float,
+            individual_size)
+        toolbox.register(
+            "population", tools.initRepeat, list, toolbox.individual)
+        toolbox.register("select", tools.selRandom, k=3)
+        toolbox.register("evaluate", self.evaluation_builder(X, y))
+
+        self.hall_of_fame = tools.HallOfFame(1)
+        stats = self._build_stats()
+
+        # TODO: Make this more general and move to BaseEvolutionStrategy
+        if self.params['population_size'] == 'log':
+            population_size = int(4 + 3 * np.log(self.params['n_dim']))
+        elif self.params['population_size'] is not None:
+            population_size = self.params['population_size']
+        else:
+            population_size = 10 * self.params['n_dim']
+        pop = toolbox.population(n=population_size)
+
+        self.logbook = tools.Logbook()
+        self.logbook.header = ['gen', 'nevals'] \
+            + (stats.fields if stats else [])
+
+        # Evaluate the individuals
+        fitnesses = toolbox.map(toolbox.evaluate, pop)
+        for ind, fit in zip(pop, fitnesses):
+            ind.fitness.values = fit
+
+        if stats:
+            record = stats.compile(pop)
+            self.logbook.record(gen=0, evals=len(pop), **record)
+            if self.params['verbose']:
+                print(self.logbook.stream)
+
+        for g in range(1, self.params['n_gen']):
+            for k, agent in enumerate(pop):
+                a, b, c = toolbox.select(pop)
+                y = toolbox.clone(agent)
+
+                # Update the control parameters
+                if np.random.random() < self.params['t1']:  # F
+                    y[0] = self.params['Fl'] \
+                        + np.random.random() * self.params['Fu']
+                if np.random.random() < self.params['t2']:  # CR
+                    y[1] = np.random.random()
+
+                # Mutation and crossover
+                index = np.random.randint(2, individual_size)
+                for i, value in enumerate(agent[2:], 2):
+                    if i == index or np.random.random() < y[1]:
+                        y[i] = a[i] + y[0] * (b[i] - c[i])
+
+                # Selection
+                y.fitness.values = toolbox.evaluate(y)
+                if y.fitness >= agent.fitness:
+                    pop[k] = y
+            self.hall_of_fame.update(pop)
+
+            if stats:
+                record = stats.compile(pop)
+                self.logbook.record(gen=g, evals=len(pop), **record)
+                if self.params['verbose']:
+                    print(self.logbook.stream)
+
+        return self
