+++ metric_learn/evolution/transformer/__init__.py
@@ -1,0 +1,6 @@
+from .diagonal import DiagonalMatrixTransformer
+from .full import FullMatrixTransformer
+# from .kmeans import KMeansTransformer
+from .matrix import MatrixTransformer
+from .neural import NeuralNetworkTransformer
+from .triangular import TriangularMatrixTransformer
