+++ test/metric_learn_test.py
@@ -7,8 +7,10 @@ from numpy.testing import assert_array_almost_equal
 
 from metric_learn import (
     LMNN, NCA, LFDA, Covariance, MLKR, MMC,
-    LSML_Supervised, ITML_Supervised, SDML_Supervised, RCA_Supervised, MMC_Supervised,
-    MetricEvolution)
+    LSML_Supervised, ITML_Supervised, SDML_Supervised, RCA_Supervised, MMC_Supervised)
+
+from metric_learn.evolution import MetricEvolution
+
 # Import this specially for testing.
 from metric_learn.lmnn import python_LMNN
